+++ priv/repo/migrations/20220328100907_add_edges_metadata_column.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddEdgesMetadataColumn do
+  use Ecto.Migration
+
+  def change do
+    alter table("edges") do
+      add(:metadata, :map)
+    end
+  end
+end